requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
apscheduler>=3.10.4
pyside6>=6.7.0
sqlalchemy>=2.0.30
deepdiff>=6.7.1
plyer>=2.1.0
tenacity>=8.2.3
tqdm>=4.66.4
python-dotenv>=1.0.1
plotly>=5.22.0